requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
